Updated most libs to get running on a Mac Pro
Installed cairo, harfbuzz, and pango (and their dependencies) from
homebrew, passing no options (i.e. --with-universal) and pulled
down the bottled (precompiled) versions. This was done on a MacBook
Pro running 10.10.5 so they should be the yosemite version.
Confirmed working on a Late 2014 MacBook Pro and 2013 Mac Pro both
running 10.10.5